raise script now works for gnome on wayland

+++ utility/raise.sh
@@ -1,3 +1,5 @@
+#!/bin/bash
+
 # Raise the passed window / start it if not running
 # 
 # param:  The process to focus / start
@@ -5,15 +7,21 @@
 # process id of this script
 PID=$$
 
-# search for the passed window / process
-xdotool search --class $1 | while read line
-do
-  # Try to activate the window
-  if [ `xdotool windowactivate $line 2> /dev/stdout | grep -c fail` -eq 0 ]
-    then
-    kill $PID
-    exit
-  fi
-done
-# launch the program if we reach here
-$1 & disown
+# Running x11 - use xdotool
+if [[ "$XDG_SESSION_TYPE" == 'x11' ]]; then
+  # search for the passed window / process
+  xdotool search --class $1 | while read line
+  do
+    # Try to activate the window
+    if [ `xdotool windowactivate $line 2> /dev/stdout | grep -c fail` -eq 0 ]
+      then
+      kill $PID
+      exit
+    fi
+  done
+# Most likely wayland: invoke gnome API using gdbus
+elif [[ "$XDG_CURRENT_DESKTOP" == 'GNOME' ]]; then
+  gdbus call --session --dest org.gnome.Shell --object-path /org/gnome/Shell --method org.gnome.Shell.Eval \
+    "global.get_window_actors().map(a=>a.meta_window).find(w=>w.get_wm_class()==\"$1\").activate(0)"
+  exit
+fi
